Move case creation related classes to one package
