+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/casecreation/CaseCreationException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.casecreation;
 
 public class CaseCreationException extends RuntimeException {
