+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/casecreation/CaseCreationResultType.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.events;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.casecreation;
 
 public enum CaseCreationResultType {
     CASE_CREATED,
